+++ src/app/hooks/useTVLSortPool.ts
@@ -1,49 +1,12 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { TokenInfo } from '@solana/spl-token-registry'
 
 import { useMint, usePool } from 'senhub/providers'
-import { account, utils } from '@senswap/sen-js'
-import { fetchCGK } from 'shared/util'
+import { getTVL } from 'app/helper'
 
 export const useTVLSortPool = (listSentrePools: string[]) => {
   const { tokenProvider } = useMint()
   const [poolTVL, setPoolTVL] = useState<Record<string, number>>({})
   const { pools } = usePool()
-
-  const getTVL = useCallback(
-    async (poolAddress: string) => {
-      try {
-        const { reserve_a, reserve_b, mint_a, mint_b } = pools[poolAddress]
-        const mintAddresses = [mint_a, mint_b]
-        const tokenInfos: Array<TokenInfo | any> = await Promise.all(
-          mintAddresses.map((mintAddress) => {
-            if (!account.isAddress(mintAddress)) return {}
-            return tokenProvider.findByAddress(mintAddress)
-          }),
-        )
-        const reserves = [reserve_a, reserve_b]
-        const decimals = tokenInfos.map(({ decimals }) => decimals)
-        const data = await Promise.all(
-          tokenInfos.map(({ extensions }) => {
-            if (!extensions?.coingeckoId) return {} as any
-            return fetchCGK(extensions?.coingeckoId)
-          }),
-        )
-        const prices = data.map(({ price }) => price)
-        let tvl = 0
-        for (const i in reserves) {
-          if (reserves[i] && decimals[i] && prices[i]) {
-            const amount = Number(utils.undecimalize(reserves[i], decimals[i]))
-            tvl = tvl + amount * prices[i]
-          }
-        }
-        return tvl
-      } catch (er) {
-        return 0
-      }
-    },
-    [pools, tokenProvider],
-  )
 
   const calcPoolTVL = useCallback(async () => {
     if (!Object.keys(pools).length) return
@@ -51,12 +14,12 @@ export const useTVLSortPool = (listSentrePools: string[]) => {
     const newPoolTVL: Record<string, number> = {}
     await Promise.all(
       Object.keys(pools).map(async (addr) => {
-        const tvl = await getTVL(addr)
+        const tvl = await getTVL(addr, tokenProvider, pools)
         newPoolTVL[addr] = tvl
       }),
     )
     setPoolTVL(newPoolTVL)
-  }, [getTVL, poolTVL, pools])
+  }, [poolTVL, pools, tokenProvider])
 
   useEffect(() => {
     calcPoolTVL()
